+++ app/controllers/DashboardController.php
@@ -1,0 +1,23 @@
+<?php
+
+class DashboardController extends \BaseController {
+
+	/**
+	 * Display a listing of the resource.
+	 *
+	 * @return Response
+	 */
+	public function index()
+	{
+		$user_id = Auth::user()->id;
+		$joined_events = JoinedEvents::with('event')->where('attendee_id', '=', $user_id)->get();
+		$hosted_events = JoinedEvents::with('event')->where('host_id', '=', $user_id)->get();
+
+		return View::make('dashboard.dashboard')
+			->with('title', 'Dashboard')
+			->with('joined_events', $joined_events)
+			->with('hosted_events', $hosted_events);
+	}
+
+
+}