+++ app/database/migrations/2015_02_01_212029_create_joinEvents_table.php
@@ -1,0 +1,35 @@
+<?php
+
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class CreateJoinEventsTable extends Migration {
+
+	/**
+	 * Run the migrations.
+	 *
+	 * @return void
+	 */
+	public function up()
+	{
+		Schema::create('joined_events', function($table){
+			$table->increments('id');
+			$table->integer('attendee_id');
+			$table->integer('host_id');
+			$table->integer('event_id');
+			$table->integer('status');
+			$table->timestamps();
+		});
+	}
+
+	/**
+	 * Reverse the migrations.
+	 *
+	 * @return void
+	 */
+	public function down()
+	{
+		Schema::drop('joined_events');
+	}
+
+}