+++ app/views/event/host.blade.php
@@ -26,9 +26,7 @@
 		<datalist id="event_type"></datalist>
 	</tr>
 	<tr>
-
-	<link rel="stylesheet" href="//code.jquery.com/ui/1.11.2/themes/smoothness/jquery-ui.css">
-	<script src="//code.jquery.com/ui/1.11.2/jquery-ui.js"></script>
+		
 	<script>
   $(function() {
     $( ".datepicker" ).datepicker({dateFormat: 'dd-mm-yy'});